Add new GitHub Actions workflow for gol2 migration
A new GitHub Actions workflow for gol2 migration has been introduced. This workflow, which runs on ubuntu-latest, covers a build process that uses docker metadata-action and docker build-push-action, and a deployment process to the DigitalOcean App Platform. Event triggers for this workflow include push to an 'update-starknet' branch or a manual workflow dispatch.