Update GitHub Actions workflow and Docker registries
The code updates the GitHub Actions workflow by including an additional branch for 'snapshot' and updating the Docker meta and build-push actions to newer versions. The Docker container registry has been shifted from DigitalOcean to Google Artifacts Registry. This change requires a new login action and the removal of the 'Deploy-review' and 'Deploy-production' steps